FIX: Lifted username retrieval from auth Token to parent ShiftForm.js from child WorkedShift.js
